refactor crd.Tagging methods into pkg/model/tagging.go
