+++ pkg/testdata/models/apis/s3/0000-00-00/generator-with-tags.yaml
@@ -25,8 +25,6 @@ resources:
         from:
           operation: PutBucketTagging
           path: Tagging
-#      Tagging.TagSet:
-#        has_tags: true
       ACL:
         # This is to test the ackcompare field ignore functionality. This
         # should NOT be in a production generator.yaml...
